Breaks out the management of the form to separate file
